+++ manual/src/release_notes.md
@@ -2,6 +2,7 @@
 
 ## 0.4.0 (WIP)
 
+- Introduce DirectSource, for music and other non-panned assets.
 - All sources now have gain, as part of refactoring for DirectSource.
 - Source add/remove generators needed to happen in the audio thread in order to not break. For now this is using `Context.call`, but in future
   it'll probably use a ConcurrentQueue to enqueue updates and avoid the overhead, or perhaps an abstracted PropertyRing.